Update npm module to work better with newer npm versions
Try to better use information from the 'npm outdated' command
in json format if available and merge it with the information
from 'npm list' to provide a more accurate view of the packages
that are installed, missing or in need of an update.

Also improved the detection of update vs install based on
existing package versions.